Refactor non-primary modules into misc/ folder
