+++ pythonturtle/misc/__init__.py
@@ -1,0 +1,3 @@
+"""
+Supportive modules and classes for PythonTurtle implementation details.
+"""
