+++ pythonturtle/misc/helpers.py
@@ -36,7 +36,8 @@ def from_resource_folder(filename):
     Absolute path for a file assumed to be in resources folder.
     """
     my_location = os.path.dirname(__file__)
-    return os.path.join(my_location, 'resources', filename)
+    package_location = os.path.dirname(my_location)
+    return os.path.join(package_location, 'resources', filename)
 
 
 def log(x):
